requirements.txt
streamlit==1.31.1
groq==0.5.0
torch==2.1.2
transformers==4.38.0
langchain_community==0.0.38
langchain==0.1.20
langchain-groq
sentence-transformers
pypdf
faiss-cpu
python-dotenv
httpx==0.27.2
PyMuPDF==1.23.14
Pillow==10.1.0
numpy==1.24.3
timm==0.9.12
